added ThemeView::view
Inheriting from ThemeView was at the time it was implemented considered
perfectly fine for all circumstances but one nasty issue that can popup in
very complex partial in partial in partial chains with multiple instances
of the same partial showing up in the same view is that you can have cases
of variable inhertance polution where inheriting from the default view
just trickeled into partials with complex variable logic and caused
unintended consequences.

While the partial implementation isn't considered to be defected by the
consequence chaining, the view method is now available for instances where
it's better to ensure all variables passed to a subview for a more
reliable/safer rendering.